+++ src/Altinn.Notifications.Core/Services/Interfaces/IStatusFeedService.cs
@@ -16,7 +16,7 @@ public interface IStatusFeedService
     /// <param name="creatorName">Name of the service owner</param>
     /// <param name="cancellationToken">A CancellationToken for cancelling an ongoing asynchronous Task</param>
     /// <returns>Result object containing, on success: a list of order status objects following the sequence number. On failure: contains a ServiceError object</returns>
-    Task<Result<List<StatusFeed>, ServiceError>> GetStatusFeed(int seq, int? pageSize, string creatorName, CancellationToken cancellationToken);
+    Task<Result<List<StatusFeed>, ServiceError>> GetStatusFeed(long seq, int? pageSize, string creatorName, CancellationToken cancellationToken);
 
     /// <summary>
     /// Deletes outdated records from the status feed table.
